tenant_portal/requirements.txt
fastapi>=0.116.0,<1.0.0
uvicorn[standard]==0.30.0
gunicorn==22.0.0
psycopg2-binary==2.9.7
pydantic==2.9.2
requests>=2.32.4,<3.0.0
httpx>=0.27.0,<1.0.0
slowapi==0.1.9
redis>=5.0.0,<6.0.0
python-jose[cryptography]>=3.4.0,<4.0.0
pyotp==2.9.0
python-multipart>=0.0.22,<1.0.0
python-dateutil==2.8.2
# Transitive dep of python-jose — pin to safe version (CVE-2026-30922)
pyasn1>=0.6.3
